Modificadas las posiciones a double de las paredes

+++ trunk/src/modelo/ParedMetal.java
@@ -12,7 +12,7 @@ Posicionable, Observable {
 	private final int ALTO = 20;
 	private final int ANCHO = 20;	
 	 
-	public ParedConcreto(int posicionEnX, int posicionEnY){
+	public ParedConcreto(double posicionEnX, double posicionEnY){
 		setX(posicionEnX);
 		setY(posicionEnY);
 		setAlto(ALTO);      //Poner la medida que se va a usar
